# docs/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bank Card API",
    "description": "API для управления банковскими картами",
    "contact": {
      "name": "Rishat",
      "email": "risatmuhamedsin@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/users/{userId}/update": {
      "put": {
        "tags": [
          "Управление пользователями"
        ],
        "summary": "Обновить данные пользователя",
        "description": "Доступно только для ADMIN. Позволяет обновить username, email или пароль пользователя.",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID пользователя",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 2
          },
          {
            "name": "updatedUserRequest",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UpdatedUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Пользователь успешно обновлён",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные"
          },
          "401": {
            "description": "Неавторизован"
          },
          "403": {
            "description": "Доступ запрещён"
          },
          "404": {
            "description": "Пользователь не найден"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/{userId}/newAdmin": {
      "put": {
        "tags": [
          "Управление пользователями"
        ],
        "summary": "Назначить пользователю роль ADMIN",
        "description": "Доступно только для ADMIN. Позволяет назначить пользователю роль администратора.",
        "operationId": "setAdminRoleToUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID пользователя",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 2
          }
        ],
        "responses": {
          "200": {
            "description": "Роль ADMIN успешно назначена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизован"
          },
          "403": {
            "description": "Доступ запрещён"
          },
          "404": {
            "description": "Пользователь не найден"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users": {
      "post": {
        "tags": [
          "Управление пользователями"
        ],
        "summary": "Получить список пользователей",
        "description": "Возвращает страницу пользователей с возможностью сортировки по username.Доступно только для ADMIN",
        "operationId": "getAllUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Номер страницы (начиная с 0)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Размер страницы",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            },
            "example": 10
          },
          {
            "name": "sortDir",
            "in": "query",
            "description": "Направление сортировки: asc или desc",
            "required": false,
            "schema": {
              "type": "string",
              "default": "asc"
            },
            "example": "asc"
          }
        ],
        "responses": {
          "200": {
            "description": "Список пользователей получен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageUserDto"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизован"
          },
          "403": {
            "description": "Доступ запрещён"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/auth/signup": {
      "post": {
        "tags": [
          "Аутентификация"
        ],
        "summary": "Регистрация пользователя",
        "operationId": "signUp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Успешная регистрация",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JwtAuthenticationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Пользователь уже существует"
          }
        }
      }
    },
    "/api/auth/signin": {
      "post": {
        "tags": [
          "Аутентификация"
        ],
        "summary": "Авторизация пользователя",
        "operationId": "singIn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Успешная аутентификация",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JwtAuthenticationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Неверные учетные данные"
          }
        }
      }
    },
    "/api/users/{userId}/delete": {
      "patch": {
        "tags": [
          "Управление пользователями"
        ],
        "summary": "Удаление пользователя",
        "description": "Доступно только для ADMIN. Блокирует пользователя по ID.",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID пользователя",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 2
          }
        ],
        "responses": {
          "200": {
            "description": "Пользователь успешно заблокирован"
          },
          "401": {
            "description": "Неавторизован"
          },
          "403": {
            "description": "Доступ запрещён"
          },
          "404": {
            "description": "Пользователь не найден"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "UpdatedUserRequest": {
        "type": "object",
        "description": "Запрос на обновление данных пользователя",
        "properties": {
          "username": {
            "type": "string",
            "description": "Имя пользователя",
            "example": "ivan_ivanov",
            "maxLength": 50,
            "minLength": 3
          },
          "email": {
            "type": "string",
            "description": "Email пользователя",
            "example": "ivan@example.com"
          },
          "password": {
            "type": "string",
            "description": "Пароль пользователя",
            "example": "StrongPass123!",
            "maxLength": 255,
            "minLength": 8
          }
        },
        "required": [
          "email",
          "username"
        ]
      },
      "UserDto": {
        "type": "object",
        "description": "Информация о пользователе",
        "properties": {
          "username": {
            "type": "string",
            "description": "Имя пользователя",
            "example": "ivan_ivanov"
          },
          "email": {
            "type": "string",
            "description": "Email пользователя",
            "example": "ivan@example.com"
          },
          "role": {
            "type": "string",
            "description": "Роль пользователя",
            "enum": [
              "ROLE_ADMIN",
              "ROLE_USER"
            ],
            "example": "USER"
          }
        },
        "required": [
          "email",
          "role",
          "username"
        ]
      },
      "PageUserDto": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDto"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "unpaged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "paged": {
            "type": "boolean"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          }
        }
      },
...
      "SignUpRequest": {
        "type": "object",
        "description": "Запрос на регистрацию",
        "properties": {
          "username": {
            "type": "string",
            "description": "Имя пользователя",
            "example": "Rishat432",
            "maxLength": 50,
            "minLength": 5
          },
          "email": {
            "type": "string",
            "description": "Адрес электронной почты",
            "example": "Rishat@example.com",
            "maxLength": 255,
            "minLength": 5
          },
          "password": {
            "type": "string",
            "description": "Пароль",
            "example": "mySuperPuperParol12314%$",
            "maxLength": 255,
            "minLength": 8
          }
        },
        "required": [
          "email",
          "password",
          "username"
        ]
      },
      "JwtAuthenticationResponse": {
        "type": "object",
        "description": "Ответ c токеном доступа",
        "properties": {
          "token": {
            "type": "string",
            "description": "Токен доступа",
            "example": "eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJhZG1pbiIsImV4cCI6MTYyMjUwNj..."
          }
        }
      },
      "SignInRequest": {
        "type": "object",
        "description": "Запрос на аутентификацию",
        "properties": {
          "username": {
            "type": "string",
            "description": "Имя пользователя",
            "example": "ADMIN ADMINOV",
            "maxLength": 50,
            "minLength": 5
          },
          "password": {
            "type": "string",
            "description": "Пароль",
            "example": "adminAdminov123$",
            "maxLength": 255,
            "minLength": 8
          }
        },
        "required": [
          "password",
          "username"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}